fix: rename encrypt util
